+++ src/main/java/de/maxhenkel/car/integration/jei/CarRecipeCategory.java
@@ -1,5 +1,4 @@
 package de.maxhenkel.car.integration.jei;
-/*
 
 import com.mojang.blaze3d.matrix.MatrixStack;
 import de.maxhenkel.car.Main;
@@ -44,7 +43,7 @@ public class CarRecipeCategory implements IRecipeCategory<CarRecipe> {
 
     @Override
     public String getTitle() {
-        return ModBlocks.CAR_WORKSHOP.func_235333_g_().getString();
+        return ModBlocks.CAR_WORKSHOP.getTranslatedName().getString();
     }
 
     @Override
@@ -82,4 +81,4 @@ public class CarRecipeCategory implements IRecipeCategory<CarRecipe> {
     public void draw(CarRecipe recipe, MatrixStack matrixStack, double mouseX, double mouseY) {
         renderer.render(matrixStack, recipe.getCar(), RECIPE_WIDTH - 30, RECIPE_HEIGHT - 54 / 4, 18);
     }
-}*/
+}