+++ src/main/java/de/maxhenkel/car/integration/jei/JEIPlugin.java
@@ -1,5 +1,4 @@
 package de.maxhenkel.car.integration.jei;
-/*
 
 import java.util.Arrays;
 import java.util.stream.Collectors;
@@ -74,4 +73,4 @@ public class JEIPlugin implements IModPlugin {
         registry.addRecipeCategories(new CarRecipeCategory(registry.getJeiHelpers().getGuiHelper()));
     }
 
-}*/
+}