+++ src/main/java/de/maxhenkel/car/integration/jei/PainterRecipeCategoryYellow.java
@@ -1,5 +1,4 @@
 package de.maxhenkel.car.integration.jei;
-/*
 
 import de.maxhenkel.car.Main;
 import de.maxhenkel.car.items.ModItems;
@@ -65,4 +64,4 @@ public class PainterRecipeCategoryYellow implements IRecipeCategory<PainterRecip
         group.set(1, new ItemStack(wrapper.getOutput()));
 
     }
-}*/
+}